kill 'manager'. just inline
